+++ packages/terra-form/src/TextField.jsx
@@ -10,15 +10,15 @@ const propTypes = {
    */
   defaultValue: PropTypes.string,
   /**
-   * Error message for when the input is invalid
+   * Error message for when the input is invalid.
    */
   error: PropTypes.node,
   /**
-   *  Help element to display with the input
+   *  Help element to display with the input.
    */
   help: PropTypes.node,
   /**
-   * Custom attributes to apply to the input
+   * Custom attributes to apply to the input.
    */
   inputAttrs: PropTypes.object,
   /**
@@ -26,27 +26,27 @@ const propTypes = {
    */
   inputId: PropTypes.string,
   /**
-   * Whether the field is inline
+   * Whether the field is inline.
    */
   isInline: PropTypes.bool,
   /**
-   * Label of the input
+   * Label of the input.
    */
   label: PropTypes.node,
   /**
-   * Attributes to attach to the label
+   * Attributes to attach to the label.
    */
   labelAttrs: PropTypes.object,
   /**
-   * Maximum number of characters user can input in this field
+   * Maximum number of characters user can input in this field.
    */
   maxLength: PropTypes.number,
   /**
-   * Minimum number of characters user must input in this field
+   * Minimum number of characters user must input in this field.
    */
   minLength: PropTypes.number,
   /**
-   * Name of the input attribute
+   * Name of the input attribute.
    */
   name: PropTypes.string,
   /**
@@ -54,11 +54,11 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * Whether the input is required
+   * Whether the input is required.
    */
   required: PropTypes.bool,
   /**
-   * Type the input will be
+   * The input type.
    */
   type: PropTypes.string,
   /**