+++ packages/terra-modal-manager/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added gainFocus and loseFocus actions to support trapping and releasing focus in the modal manager.
+
+1.2.0 - (July 26, 2017)
+------------------
 ### Changed
 * Updated spacing around react-docgen comments